feat: added reminders, changed snowflake IDs from text to bigint, segmented frontend into components

+++ migrations/20260527135738_temp_bans.sql
@@ -1,7 +1,8 @@
-CREATE TABLE IF NOT EXISTS temp_bans (
-    id SERIAL PRIMARY KEY,
-    guild_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+CREATE TABLE IF NOT EXISTS temp_bans
+(
+    id       BIGSERIAL PRIMARY KEY,
+    guild_id BIGINT      NOT NULL,
+    user_id  BIGINT      NOT NULL,
     unban_at TIMESTAMPTZ NOT NULL
 );
